fix: Use shared button component for consistent look/feel
Use the shared `app_button` to keep everything looking same same 😉
